Move ClusterServiceConfig to internal shared directory
